Added new directory dtos with two empty import dtos. Added directory entities in domain directory and moved entities there. Solution for Make better Structure task.
