+++ src/main/java/com/emailspringproject/emailholder/controllers/EmailController.java
@@ -1,7 +1,7 @@
 package com.emailspringproject.emailholder.controllers;
 
 
-import com.emailspringproject.emailholder.domain.Email;
+import com.emailspringproject.emailholder.domain.entities.Email;
 import com.emailspringproject.emailholder.repositories.EmailRepository;
 import com.emailspringproject.emailholder.repositories.SiteRepository;
 import com.emailspringproject.emailholder.services.EmailService;
