+++ src/main/java/com/emailspringproject/emailholder/domain/dtos/EmailImportDTO.java
@@ -1,0 +1,4 @@
+package com.emailspringproject.emailholder.domain.dtos;
+
+public class EmailImportDTO {
+}
