+++ src/main/java/com/emailspringproject/emailholder/domain/entities/Email.java
@@ -1,4 +1,4 @@
-package com.emailspringproject.emailholder.domain;
+package com.emailspringproject.emailholder.domain.entities;
 
 import jakarta.persistence.*;
