+++ src/main/java/com/emailspringproject/emailholder/repositories/EmailRepository.java
@@ -1,6 +1,6 @@
 package com.emailspringproject.emailholder.repositories;
 
-import com.emailspringproject.emailholder.domain.Email;
+import com.emailspringproject.emailholder.domain.entities.Email;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmailRepository extends JpaRepository<Email, Long> {
