+++ src/main/java/com/emailspringproject/emailholder/repositories/SiteRepository.java
@@ -1,6 +1,6 @@
 package com.emailspringproject.emailholder.repositories;
 
-import com.emailspringproject.emailholder.domain.Site;
+import com.emailspringproject.emailholder.domain.entities.Site;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SiteRepository extends JpaRepository<Site, Long> {
