+++ src/main/java/com/emailspringproject/emailholder/services/impl/SiteServiceImpl.java
@@ -1,6 +1,6 @@
 package com.emailspringproject.emailholder.services.impl;
 
-import com.emailspringproject.emailholder.domain.Site;
+import com.emailspringproject.emailholder.domain.entities.Site;
 import com.emailspringproject.emailholder.repositories.EmailRepository;
 import com.emailspringproject.emailholder.repositories.SiteRepository;
 import com.emailspringproject.emailholder.services.SiteService;
